+++ src/routing/createRouters.ts
@@ -1,0 +1,11 @@
+import teapot from "./routers/teapot"
+import artist from "./routers/artist";
+
+
+
+export default function makeRouterInfos() {
+    return [
+        teapot(),
+        artist()
+    ];
+}
